refactor: Improving restrictions for defaults values
